requirements.txt
pyqt5==5.15.9
pandas==1.5.3
albumentations==1.3.1
paddlepaddle>=2.5.0
paddlepaddle-gpu>=2.5.0
paddleocr>=2.7.0
torch==2.0.1
torchvision==0.15.2
requests==2.31.0
pyautogui==0.9.54
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
pynput==1.7.6
screeninfo==0.8.1
pywin32==306
python-dotenv==1.0.0
clip @ git+https://github.com/openai/CLIP.git